Add files via upload
To run the aglorithms, make sure your location in command prompt(or any other command-line interpreter) is in the directory which contains the files.
e.g:    python linearregression.py
The whole point of both algorithms is to predict the value of a variable based on the value of another variable. This is possible with the help of our loss function, in our case our own defined function called "compute_mse()", hyperparameters as epochs and learning rate and computing gradients and optimizing weights.